New function: xfce_titled_dialog_new_with_mixed_buttons
Using xfce_titled_dialog_new_with_buttons in Gtk3 results in
buttons	that cannot have icons except for STOCK	icons. This
is because it just calls gtk_dialog_add_button to create
and add	the buttons. With this new function, you specify
an icon name, STOCK icon, or simply "" for no icon on the
button followed by the button text and response id for each
button.